Enforce ordering by replication key during full refresh
If we stop an initial data load midway our state will be that of the last record that was written to the target. If we don't enforce the ordering of the records from the source we have no guarantee that all records before the last one were written to the target (as Postgres might return them in disorder). To overcome these issues we should enforce ordering.